feat(workflow): migrate the wait node to the catalog
Phase 1 step 2 - the template migration that sets the shape for the other 37
node types. Zero diff under wait's node.tsx / panel.tsx / trace-renderer.tsx.

The data half (WaitType, DurationUnit, WaitNodeData, the zod schema, the
defaults, the validator) moves to catalog/nodes/wait.ts as the single
source. The base node data shape and its zod schema move with it
(catalog/node-base.ts) since every node schema extends them; the web
node-base.ts keeps the React Flow types and narrows `type` back to the
NodeType enum, and the web wait types do the same over the catalog's data
interface. core/wait/schema.ts shrinks to the merge site: defineFromManifest
plus the web-only output resolver plus back-compat re-exports, so neither
the panel, core/index.ts, nor the parity suite's node-definitions.ts changes
an import.

Shadow-declaration sweep for this type: the engine's nodes/wait/types.ts
re-exports the catalog enums instead of re-declaring them, keeping only the
sequence-compiled extensions (deliveryWindow, anchor) that the builder never
authors - documented as deliberately outside the manifest schema.

The registry becomes a static explicit-list map (no registration side
effects, import order can never change what is registered; registerManifest
is gone). 'wait' leaves NOT_YET_MIGRATED, and the coverage test's
defaults-parse fixture now supplies factory-assigned identity (id/type),
which is the one legitimate gap in any node's defaults.
